feat: Add authentication check on homepage
This commit adds an authentication check on the homepage to determine if the user is registered or not. If the user is not registered, they will be redirected to the registration page. The authentication check is performed using the `checkAuthToken` function from the `HomePage` component. If the authentication is successful, the user will be able to access the homepage. Otherwise, they will be redirected to the registration page.

Resolves: #123